+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.12] - 2023-07-12
+### Added
+- Add EnumHelper GetToggledFlagsCount & Clamp IComparable functions.
+
 ## [1.0.11] - 2023-07-07
 ### Added
 - Add mathmod function.